fix: revert change to ArgumentMultimap::put
Temporarily reverting the change as it was causing the AB3 test cases to fail.

This was initially changed so that arguments like ingredients and steps
can be given as one input by using "," as a delimiter. When implementing
that feature (add/edit) we'll have to either modify it for our use later
or add another method specifically for CSV values.