test: include more test and refactor test name
I have also included a test to ensure that the data in the predicate is
immutable.

As mentioned by Aidan, I've renamed the test to make it easier to
distinguish what the tests do. Building off from his suggestion, I
recommend that the tests methods would be named after the method that we
are testing try not to group tests together (split up equals check and
hashCode check).